add helper for common tagging (#1015)
Adds a sub-project with a simple helper for extracting
Netflix common tags from the environment. This is using
the same logic as `iep-nflxenv` which will later get
updated to use this new package. The goal for this
change is to remove any dependency from spectator on
iep libraries that will likely move to require java 17
in the near future while spectator will need to stay
on java 8 for the time being.