reduce warning volume for sidecar registry (#1020)
Some writers such as the UDP writer can be quite noisy if
the sidecar is not present. To avoid spamming the user with
warnings, they will be suppressed after a warning is logged.
Warnings will be re-enabled if there is a successful write
so that if there are problems with the sidecar being
intermittently accessible it will still get detected.